chore(replay): remove REPLAY_SNAPSHOT_STORE flag, simplify to store-only loading, restore post-processing
Remove the REPLAY_SNAPSHOT_STORE feature flag and all legacy loading
paths (snapshotsBySource cache, processAllSnapshots pipeline, legacy
buffer-ahead scheduling). The SnapshotStore + LoadingScheduler is now
the only snapshot loading path.

Restore the 5 post-processing behaviors that were lost when the store
path replaced processAllSnapshots: hash-based deduplication, chrome
extension stripping, mobile meta-event patching, main-thread yielding,
and analytics captures. These now run in postProcessSnapshots() inside
snapshotDataLogic's loader, with cross-source state persisted on cache.
