+++ frontend/src/scenes/session-recordings/player/sessionRecordingMetaLogic.ts
@@ -45,7 +45,6 @@ export const sessionRecordingMetaLogic = kea<sessionRecordingMetaLogicType>([
                 snapshotLogic,
                 [
                     'snapshotSources',
-                    'snapshotsBySources',
                     'snapshotsLoading',
                     'snapshotsLoaded',
                     'isLoadingSnapshots',
