+++ frontend/src/scenes/session-recordings/player/snapshot-processing/process-all-snapshots.ts
@@ -1,29 +1,15 @@
 import posthog, { PostHog } from 'posthog-js'
 
-import { EventType, eventWithTime, fullSnapshotEvent } from '@posthog/rrweb-types'
+import { eventWithTime } from '@posthog/rrweb-types'
 
-import { isEmptyObject, isObject } from 'lib/utils'
+import { isObject } from 'lib/utils'
 import { transformEventToWeb } from 'scenes/session-recordings/mobile-replay'
-import {
-    CHROME_EXTENSION_DENY_LIST,
-    stripChromeExtensionDataFromNode,
-} from 'scenes/session-recordings/player/snapshot-processing/chrome-extension-stripping'
 import { chunkMutationSnapshot } from 'scenes/session-recordings/player/snapshot-processing/chunk-large-mutations'
 import { decompressEvent } from 'scenes/session-recordings/player/snapshot-processing/decompress'
 import { getDecompressionWorkerManager } from 'scenes/session-recordings/player/snapshot-processing/DecompressionWorkerManager'
-import {
-    ViewportResolution,
-    extractDimensionsFromMobileSnapshot,
-} from 'scenes/session-recordings/player/snapshot-processing/patch-meta-event'
-import { SourceKey, keyForSource } from 'scenes/session-recordings/player/snapshot-processing/source-key'
 import { throttleCapture } from 'scenes/session-recordings/player/snapshot-processing/throttle-capturing'
 
-import {
-    EncodedRecordingSnapshot,
-    RecordingSnapshot,
-    SessionRecordingSnapshotSource,
-    SessionRecordingSnapshotSourceResponse,
-} from '~/types'
+import { EncodedRecordingSnapshot, RecordingSnapshot } from '~/types'
 
 export type RegisterWindowIdCallback = (uuid: string) => number
 
@@ -39,318 +25,6 @@ export const createWindowIdRegistry = (): RegisterWindowIdCallback => {
     }
 }
 
-export type ProcessingCache = {
-    snapshots: Record<SourceKey, RecordingSnapshot[]>
-}
-
-function extractImgNodeFromMobileIncremental(snapshot: RecordingSnapshot): any | undefined {
-    if (snapshot.type !== EventType.IncrementalSnapshot) {
-        return undefined
-    }
-    const data: any = (snapshot as any).data
-
-    if (data?.source !== 0 || !Array.isArray(data.adds)) {
-        return undefined
-    }
-
-    const checksLimit = Math.min(data.adds.length, 3)
-    for (let i = 0; i < checksLimit; i++) {
-        const node = data.adds[i]?.node
-        if (
-            node &&
-            node.type === 2 &&
-            node.tagName === 'img' &&
-            node.attributes?.['data-rrweb-id'] &&
-            node.attributes?.width &&
-            node.attributes?.height
-        ) {
-            return node
-        }
-    }
-
-    return undefined
-}
-
-function isLikelyMobileScreenshot(snapshot: RecordingSnapshot): boolean {
-    return extractImgNodeFromMobileIncremental(snapshot) !== undefined
-}
-
-function createMinimalFullSnapshot(windowId: number | undefined, timestamp: number, imgNode?: any): RecordingSnapshot {
-    // Create a minimal rrweb full document snapshot structure sufficient for playback
-    // For mobile screenshots, include the img node in body so dimension extraction works
-    const bodyChildNodes = imgNode ? [imgNode] : []
-
-    const htmlNode = {
-        type: 2, // Element node
-        tagName: 'html',
-        attributes: {
-            'data-rrweb-id': 'minimal-html',
-        },
-        childNodes: [
-            {
-                type: 2, // Element node
-                tagName: 'head',
-                attributes: {
-                    'data-rrweb-id': 'minimal-head',
-                },
-                childNodes: [],
-            },
-            {
-                type: 2, // Element node
-                tagName: 'body',
-                attributes: {
-                    'data-rrweb-id': 5, // Match mobile transformer body id
-                },
-                childNodes: bodyChildNodes,
-            },
-        ],
-    }
-    const documentNode = {
-        type: 0, // NodeType.Document
-        childNodes: [htmlNode],
-    }
-    return {
-        type: EventType.FullSnapshot,
-        timestamp,
-        windowId,
-        data: {
-            node: documentNode,
-            initialOffset: { top: 0, left: 0 },
-        },
-    } as unknown as RecordingSnapshot
-}
-
-export async function processAllSnapshots(
-    sources: SessionRecordingSnapshotSource[] | null,
-    snapshotsBySource: Record<SourceKey, SessionRecordingSnapshotSourceResponse> | null,
-    processingCache: ProcessingCache,
-    viewportForTimestamp: (timestamp: number) => ViewportResolution | undefined,
-    sessionRecordingId: string
-): Promise<RecordingSnapshot[]> {
-    if (!sources || !snapshotsBySource || isEmptyObject(snapshotsBySource)) {
-        return []
-    }
-
-    const context: ProcessSnapshotContext = {
-        result: [],
-        sourceResult: [],
-        matchedExtensions: new Set<string>(),
-        hasSeenMeta: false,
-        seenFullByWindow: {},
-        previousTimestamp: null,
-        seenHashes: new Set<number>(),
-    }
-
-    const YIELD_AFTER_MS = 50
-    let lastYield = performance.now()
-
-    for (let sourceIdx = 0; sourceIdx < sources.length; sourceIdx++) {
-        const source = sources[sourceIdx]
-        const sourceKey = keyForSource(source)
-
-        if (sourceKey in processingCache.snapshots) {
-            const cachedSnapshots = processingCache.snapshots[sourceKey]
-            for (const snapshot of cachedSnapshots) {
-                context.result.push(snapshot)
-            }
-            continue
-        }
-
-        if (!(sourceKey in snapshotsBySource)) {
-            continue
-        }
-
-        const sourceSnapshots = snapshotsBySource[sourceKey].snapshots || []
-        if (!sourceSnapshots.length) {
-            continue
-        }
-
-        context.sourceResult = []
-        const sortedSnapshots = sourceSnapshots.sort((a, b) => a.timestamp - b.timestamp)
-        context.seenHashes = new Set<number>()
-        const pushPatchedMeta = createPushPatchedMeta(context, sourceKey, viewportForTimestamp, sessionRecordingId)
-
-        for (let snapshotIndex = 0; snapshotIndex < sortedSnapshots.length; snapshotIndex++) {
-            const snapshot = sortedSnapshots[snapshotIndex]
-            processSnapshot(
-                snapshot,
-                sortedSnapshots,
-                snapshotIndex,
-                context,
-                pushPatchedMeta,
-                sessionRecordingId,
-                sourceKey
-            )
-            if (performance.now() - lastYield > YIELD_AFTER_MS) {
-                await new Promise<void>((r) => setTimeout(r, 0))
-                lastYield = performance.now()
-            }
-        }
-
-        processingCache.snapshots[sourceKey] = context.sourceResult
-
-        if (snapshotsBySource[sourceKey]) {
-            snapshotsBySource[sourceKey].snapshots = []
-        }
-    }
-
-    context.result.sort((a, b) => a.timestamp - b.timestamp)
-
-    return context.result
-}
-
-type ProcessSnapshotContext = {
-    result: RecordingSnapshot[]
-    sourceResult: RecordingSnapshot[]
-    matchedExtensions: Set<string>
-    hasSeenMeta: boolean
-    seenFullByWindow: Record<number, boolean>
-    previousTimestamp: number | null
-    seenHashes: Set<number>
-}
-
-function createPushPatchedMeta(
-    context: ProcessSnapshotContext,
-    sourceKey: SourceKey,
-    viewportForTimestamp: (timestamp: number) => ViewportResolution | undefined,
-    sessionRecordingId: string
-): (ts: number, winId?: number, fullSnapshot?: RecordingSnapshot) => boolean {
-    return (ts: number, winId?: number, fullSnapshot?: RecordingSnapshot): boolean => {
-        if (context.hasSeenMeta) {
-            return false
-        }
-
-        let viewport: ViewportResolution | undefined
-        if (fullSnapshot) {
-            viewport = extractDimensionsFromMobileSnapshot(fullSnapshot)
-        }
-
-        if (!viewport) {
-            viewport = viewportForTimestamp(ts)
-        }
-
-        if (viewport && viewport.width && viewport.height) {
-            const metaEvent: RecordingSnapshot = {
-                type: EventType.Meta,
-                timestamp: ts,
-                windowId: winId as number,
-                data: {
-                    width: parseInt(viewport.width, 10),
-                    height: parseInt(viewport.height, 10),
-                    href: viewport.href || 'unknown',
-                },
-            }
-            context.result.push(metaEvent)
-            context.sourceResult.push(metaEvent)
-            throttleCapture(`${sessionRecordingId}-patched-meta`, () => {
-                posthog.capture('patched meta into web recording', {
-                    throttleCaptureKey: `${sessionRecordingId}-patched-meta`,
-                    sessionRecordingId,
-                    sourceKey: sourceKey,
-                    feature: 'session-recording-meta-patching',
-                })
-            })
-            return true
-        }
-        throttleCapture(`${sessionRecordingId}-no-viewport-found`, () => {
-            posthog.captureException(new Error('No event viewport or meta snapshot found for full snapshot'), {
-                throttleCaptureKey: `${sessionRecordingId}-no-viewport-found`,
-                sessionRecordingId,
-                sourceKey: sourceKey,
-                feature: 'session-recording-meta-patching',
-            })
-        })
-        return false
-    }
-}
-
-function processSnapshot(
-    snapshot: RecordingSnapshot,
-    sortedSnapshots: RecordingSnapshot[],
-    snapshotIndex: number,
-    context: ProcessSnapshotContext,
-    pushPatchedMeta: (ts: number, winId?: number, fullSnapshot?: RecordingSnapshot) => boolean,
-    sessionRecordingId: string,
-    sourceKey: SourceKey
-): void {
-    const currentTimestamp = snapshot.timestamp
-
-    if (currentTimestamp === context.previousTimestamp) {
-        if (context.seenHashes.size === 0) {
-            context.seenHashes.add(hashSnapshot(sortedSnapshots[snapshotIndex - 1]))
-        }
-        const snapshotHash = hashSnapshot(snapshot)
-        if (!context.seenHashes.has(snapshotHash)) {
-            context.seenHashes.add(snapshotHash)
-        } else {
-            throttleCapture(`${sessionRecordingId}-duplicate-snapshot`, () => {
-                posthog.capture('session recording has duplicate snapshots', {
-                    sessionRecordingId,
-                    sourceKey: sourceKey,
-                })
-            })
-            return
-        }
-    } else {
-        context.seenHashes = new Set<number>()
-    }
-
-    if (snapshot.type === EventType.Meta) {
-        context.hasSeenMeta = true
-    }
-
-    const windowId = snapshot.windowId
-    const hasSeenFullForWindow = !!context.seenFullByWindow[windowId]
-
-    if (snapshot.type === EventType.IncrementalSnapshot && isLikelyMobileScreenshot(snapshot)) {
-        const syntheticTimestamp = Math.max(0, snapshot.timestamp - 1)
-        const imgNode = extractImgNodeFromMobileIncremental(snapshot)
-        const syntheticFull = createMinimalFullSnapshot(snapshot.windowId, syntheticTimestamp, imgNode)
-
-        // Only patch meta if we haven't seen a full snapshot for this window yet
-        if (!hasSeenFullForWindow) {
-            const metaInserted = pushPatchedMeta(syntheticTimestamp, snapshot.windowId, syntheticFull)
-            context.hasSeenMeta = context.hasSeenMeta || metaInserted
-            context.seenFullByWindow[windowId] = true
-        }
-
-        context.result.push(syntheticFull)
-        context.sourceResult.push(syntheticFull)
-    }
-
-    if (snapshot.type === EventType.FullSnapshot) {
-        context.seenFullByWindow[snapshot.windowId] = true
-        pushPatchedMeta(snapshot.timestamp, snapshot.windowId, snapshot)
-        context.hasSeenMeta = false
-
-        const fullSnapshot = snapshot as RecordingSnapshot & fullSnapshotEvent & eventWithTime
-
-        if (
-            stripChromeExtensionDataFromNode(
-                fullSnapshot.data.node,
-                Object.keys(CHROME_EXTENSION_DENY_LIST),
-                context.matchedExtensions
-            )
-        ) {
-            context.result.unshift({
-                type: EventType.Custom,
-                data: {
-                    tag: 'chrome-extension-stripped',
-                    payload: {
-                        extensions: Array.from(context.matchedExtensions),
-                    },
-                },
-                timestamp: snapshot.timestamp,
-                windowId: snapshot.windowId,
-            })
-        }
-    }
-
-    context.result.push(snapshot)
-    context.sourceResult.push(snapshot)
-    context.previousTimestamp = currentTimestamp
-}
-
 function isRecordingSnapshot(x: unknown): x is RecordingSnapshot {
     return (
         typeof x === 'object' &&
@@ -364,7 +38,6 @@ function isRecordingSnapshot(x: unknown): x is RecordingSnapshot {
 
 const mobileFullSnapshot = (x: Record<string, any>): boolean => isObject(x.data) && 'wireframes' in x.data
 
-// the mobileFullSnapshot above wasn't catching recordings from React Native SDK 4.1.0 that were missing meta events so...
 const mobileIncrementalUpdate = (y: Record<string, any>): boolean => {
     return (
         'type' in y &&
@@ -382,19 +55,7 @@ export function hasAnyWireframes(snapshotData: Record<string, any>[]): boolean {
     })
 }
 
-function hashSnapshot(snapshot: RecordingSnapshot): number {
-    const { delay, ...delayFreeSnapshot } = snapshot
-    return cyrb53(JSON.stringify(delayFreeSnapshot))
-}
-
-/**
- * We can receive data in one of multiple formats, so we treat it as unknown,
- * And if we can't process it, force it into eventWithTime
- *
- * If it can't be case as eventWithTime by this point, then it's probably not a valid event anyway
- */
 function coerceToEventWithTime(d: unknown, sessionRecordingId: string): eventWithTime {
-    // we decompress first so that we could support partial compression on mobile in the future
     const currentEvent = decompressEvent(d, sessionRecordingId)
     return transformEventToWeb(currentEvent) ?? (currentEvent as eventWithTime)
 }
@@ -420,12 +81,10 @@ function isLengthPrefixedSnappy(uint8Data: Uint8Array): boolean {
 const lengthPrefixedSnappyDecompress = async (uint8Data: Uint8Array, posthogInstance?: PostHog): Promise<string> => {
     const workerManager = getDecompressionWorkerManager(posthogInstance)
 
-    // Phase 1: Parse and collect all compressed blocks
     const compressedBlocks: Uint8Array[] = []
     let offset = 0
 
     while (offset < uint8Data.byteLength) {
-        // Read 4-byte length prefix (big-endian unsigned int)
         if (offset + 4 > uint8Data.byteLength) {
             console.error('Incomplete length prefix at offset', offset)
             break
@@ -439,7 +98,6 @@ const lengthPrefixedSnappyDecompress = async (uint8Data: Uint8Array, posthogInst
             0
         offset += 4
 
-        // Read compressed block
         if (offset + length > uint8Data.byteLength) {
             console.error(
                 `Incomplete block at offset ${offset}, expected ${length} bytes, available ${uint8Data.byteLength - offset}`
@@ -447,15 +105,11 @@ const lengthPrefixedSnappyDecompress = async (uint8Data: Uint8Array, posthogInst
             break
         }
 
-        // Create a copy of the block to avoid ArrayBuffer detachment issues
-        // when transferring to workers in parallel
         const compressedBlock = uint8Data.slice(offset, offset + length)
         compressedBlocks.push(compressedBlock)
         offset += length
     }
 
-    // Phase 2: Decompress blocks in batches, yielding between batches to avoid
-    // microtask storms that block the main thread when many blocks resolve at once
     const isParallel = compressedBlocks.length > 1
     const DECOMPRESSION_BATCH_SIZE = 10
     const decompressedBlocks: Uint8Array[] = []
@@ -468,7 +122,6 @@ const lengthPrefixedSnappyDecompress = async (uint8Data: Uint8Array, posthogInst
         }
     }
 
-    // Phase 3: Decode all blocks to strings
     const textDecoder = new TextDecoder('utf-8')
     const decompressedParts = decompressedBlocks.map((data) => textDecoder.decode(data))
 
@@ -513,7 +166,6 @@ export const parseEncodedSnapshots = async (
 
     const registerFn: RegisterWindowIdCallback = registerWindowId || createWindowIdRegistry()
 
-    // Check if we received binary data (ArrayBuffer or Uint8Array)
     if (items instanceof ArrayBuffer || items instanceof Uint8Array) {
         const uint8Data = items instanceof Uint8Array ? items : new Uint8Array(items)
 
@@ -661,25 +313,4 @@ export const parseEncodedSnapshots = async (
     }
     reportUnparseableLines()
     return parsedLines
-}
-
-/*
-    cyrb53 (c) 2018 bryc (github.com/bryc)
-    License: Public domain. Attribution appreciated.
-    A fast and simple 53-bit string hash function with decent collision resistance.
-    Largely inspired by MurmurHash2/3, but with a focus on speed/simplicity.
-*/
-const cyrb53 = function (str: string, seed = 0): number {
-    let h1 = 0xdeadbeef ^ seed,
-        h2 = 0x41c6ce57 ^ seed
-    for (let i = 0, ch; i < str.length; i++) {
-        ch = str.charCodeAt(i)
-        h1 = Math.imul(h1 ^ ch, 2654435761)
-        h2 = Math.imul(h2 ^ ch, 1597334677)
-    }
-    h1 = Math.imul(h1 ^ (h1 >>> 16), 2246822507)
-    h1 ^= Math.imul(h2 ^ (h2 >>> 13), 3266489909)
-    h2 = Math.imul(h2 ^ (h2 >>> 16), 2246822507)
-    h2 ^= Math.imul(h1 ^ (h1 >>> 13), 3266489909)
-    return 4294967296 * (2097151 & h2) + (h1 >>> 0)
 }
