+++ frontend/src/scenes/session-recordings/player/snapshotDataLogic.tsx
@@ -5,11 +5,14 @@ import { loaders } from 'kea-loaders'
 import posthog from 'posthog-js'
 
 import api, { RecordingDeletedError } from 'lib/api'
-import { FEATURE_FLAGS } from 'lib/constants'
 import 'lib/dayjs'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
+import { ViewportResolution } from 'scenes/session-recordings/player/snapshot-processing/patch-meta-event'
+import {
+    createPostProcessingState,
+    postProcessSnapshots,
+} from 'scenes/session-recordings/player/snapshot-processing/post-process-snapshots'
 import { parseEncodedSnapshots } from 'scenes/session-recordings/player/snapshot-processing/process-all-snapshots'
-import { SourceKey, keyForSource } from 'scenes/session-recordings/player/snapshot-processing/source-key'
+import { keyForSource } from 'scenes/session-recordings/player/snapshot-processing/source-key'
 import { windowIdRegistryLogic } from 'scenes/session-recordings/player/windowIdRegistryLogic'
 
 import {
@@ -46,8 +49,6 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
         values: [
             windowIdRegistryLogic({ sessionRecordingId: props.sessionRecordingId }),
             ['uuidToIndex', 'getWindowId'],
-            featureFlagLogic,
-            ['featureFlags'],
         ],
     })),
     actions({
@@ -67,6 +68,7 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
         updatePlaybackPosition: (timestamp: number) => ({ timestamp }),
         setPlayerActive: (active: boolean) => ({ active }),
         loadAllSources: true,
+        setViewportCallback: (fn: ((ts: number) => ViewportResolution | undefined) | null) => ({ fn }),
     }),
     reducers(() => ({
         snapshotsBySourceSuccessCount: [
@@ -205,20 +207,63 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
                             actions.registerWindowId(uuid)
                         }
                     }
-                    if (cache.useSnapshotStore) {
-                        cache.pendingBatch = { sources, snapshots: parsedSnapshots }
-                    } else {
-                        // Legacy path: accumulate in cache.snapshotsBySource
-                        if (!cache.snapshotsBySource) {
-                            cache.snapshotsBySource = {}
+
+                    // Post-process: dedup, chrome extension stripping, mobile meta patching
+                    if (!cache.postProcessingState) {
+                        cache.postProcessingState = createPostProcessingState()
+                    }
+                    const processedSnapshots = await postProcessSnapshots(
+                        parsedSnapshots,
+                        cache.postProcessingState,
+                        props.sessionRecordingId,
+                        cache.viewportForTimestamp
+                    )
+
+                    // Populate the store BEFORE the SUCCESS action fires.
+                    // storeVersion recomputes from snapshotsBySourceSuccessCount
+                    // (a reducer on the SUCCESS action). If markLoaded happens in
+                    // the SUCCESS listener instead, the reducer fires first, the
+                    // selector reads a stale store version, and downstream selectors
+                    // (snapshots, fullyLoaded) see empty data.
+                    const allSources = values.snapshotSources
+                    if (allSources && cache.store) {
+                        const sourceEntries: { sourceIndex: number; endMs: number }[] = []
+                        for (const loaded of sources) {
+                            const sourceIndex = allSources.findIndex((s) => keyForSource(s) === keyForSource(loaded))
+                            if (sourceIndex >= 0) {
+                                const entry = cache.store.getEntry(sourceIndex)
+                                if (entry) {
+                                    sourceEntries.push({ sourceIndex, endMs: entry.endMs })
+                                }
+                            }
                         }
-                        const storageKey = keyForSource(sources[0])
-                        cache.snapshotsBySource[storageKey] = { snapshots: parsedSnapshots }
-                        sources.forEach((s) => {
-                            const k = keyForSource(s)
-                            cache.snapshotsBySource[k] = cache.snapshotsBySource[k] || {}
-                            cache.snapshotsBySource[k].sourceLoaded = true
-                        })
+                        sourceEntries.sort((a, b) => a.endMs - b.endMs)
+
+                        const buckets = new Map<number, RecordingSnapshot[]>()
+                        for (const se of sourceEntries) {
+                            buckets.set(se.sourceIndex, [])
+                        }
+                        let seIdx = 0
+                        for (const snap of processedSnapshots) {
+                            if (sourceEntries.length === 0) {
+                                break
+                            }
+                            while (seIdx < sourceEntries.length - 1 && snap.timestamp > sourceEntries[seIdx].endMs) {
+                                seIdx++
+                            }
+                            buckets.get(sourceEntries[seIdx].sourceIndex)?.push(snap)
+                        }
+                        for (const se of sourceEntries) {
+                            cache.store.markLoaded(se.sourceIndex, buckets.get(se.sourceIndex)!)
+                        }
+
+                        if (
+                            !processedSnapshots.length &&
+                            allSources.length === 1 &&
+                            allSources[0].source !== SnapshotSourceType.file
+                        ) {
+                            posthog.capture('recording_snapshots_v2_empty_response', { source: allSources[0] })
+                        }
                     }
 
                     return { sources }
@@ -227,8 +272,12 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
         ],
     })),
     listeners(({ values, actions, cache, props }) => ({
+        setViewportCallback: ({ fn }) => {
+            cache.viewportForTimestamp = fn ?? undefined
+        },
+
         setTargetTimestamp: ({ timestamp }) => {
-            if (!cache.useSnapshotStore || !cache.scheduler || !cache.store) {
+            if (!cache.scheduler || !cache.store) {
                 return
             }
             if (timestamp !== null) {
@@ -264,9 +313,6 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
         },
 
         updatePlaybackPosition: ({ timestamp }) => {
-            if (!cache.useSnapshotStore) {
-                return
-            }
             cache.playbackPosition = timestamp
             // Trigger loading if the buffer ahead needs filling
             actions.loadNextSnapshotSource()
@@ -274,36 +320,22 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
 
         setPlayerActive: ({ active }) => {
             cache.playerActive = active
-            if (active && cache.useSnapshotStore) {
+            if (active) {
                 actions.loadNextSnapshotSource()
             }
         },
 
         loadAllSources: () => {
-            if (cache.useSnapshotStore && cache.scheduler) {
-                cache.scheduler.loadAll()
-                actions.loadNextSnapshotSource()
-            }
+            cache.scheduler.loadAll()
+            actions.loadNextSnapshotSource()
         },
 
         setSnapshots: ({ snapshots }: { snapshots: RecordingSnapshot[] }) => {
-            if (cache.useSnapshotStore && cache.store) {
-                const fileSource = { source: SnapshotSourceType.file } as SessionRecordingSnapshotSource
-                cache.store.setSources([fileSource])
-                cache.store.markLoaded(0, snapshots)
-                cache.pendingBatch = { snapshots, sources: [fileSource] }
-            } else {
-                cache.snapshotsBySource = {
-                    'file-file': {
-                        snapshots: snapshots,
-                        source: { source: SnapshotSourceType.file },
-                        sourceLoaded: true,
-                    },
-                }
-            }
+            const fileSource = { source: SnapshotSourceType.file } as SessionRecordingSnapshotSource
+            cache.store.setSources([fileSource])
+            cache.store.markLoaded(0, snapshots)
 
             // Set sources first, then trigger the success action
-            // Otherwise processSnapshotsAsync will see null sources
             actions.loadSnapshotSourcesSuccess([{ source: SnapshotSourceType.file }])
             actions.loadSnapshotsForSourceSuccess({
                 source: { source: SnapshotSourceType.file },
@@ -340,82 +372,15 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
                 actions.setPollingInterval(newInterval)
             }
 
-            if (cache.useSnapshotStore && sourcesChanged) {
+            if (sourcesChanged) {
                 cache.store.setSources(snapshotSources)
             }
 
             actions.loadNextSnapshotSource()
         },
 
-        loadSnapshotsForSourceSuccess: ({ snapshotsForSource }) => {
+        loadSnapshotsForSourceSuccess: () => {
             cache.loadFailureCount = 0
-            const sources = values.snapshotSources
-
-            if (cache.useSnapshotStore && sources) {
-                // Store path: read from pendingBatch, bucket into SnapshotStore
-                const pending = cache.pendingBatch
-                cache.pendingBatch = null
-                const allBatchSnapshots: RecordingSnapshot[] = pending?.snapshots || []
-                const loadedSources: Pick<SessionRecordingSnapshotSource, 'source' | 'blob_key'>[] = pending?.sources ||
-                    snapshotsForSource.sources || [snapshotsForSource.source]
-
-                if (
-                    !allBatchSnapshots.length &&
-                    sources.length === 1 &&
-                    sources[0].source !== SnapshotSourceType.file
-                ) {
-                    posthog.capture('recording_snapshots_v2_empty_response', { source: sources[0] })
-                }
-
-                // Build ordered list of (sourceIndex, endMs) for the loaded sources
-                const sourceEntries: { sourceIndex: number; endMs: number }[] = []
-                for (const loaded of loadedSources) {
-                    const sourceIndex = sources.findIndex((s) => keyForSource(s) === keyForSource(loaded))
-                    if (sourceIndex < 0) {
-                        continue
-                    }
-                    const entry = cache.store.getEntry(sourceIndex)
-                    if (!entry) {
-                        continue
-                    }
-                    sourceEntries.push({ sourceIndex, endMs: entry.endMs })
-                }
-                sourceEntries.sort((a, b) => a.endMs - b.endMs)
-
-                // Split snapshots across sources by timestamp range
-                const buckets = new Map<number, RecordingSnapshot[]>()
-                for (const se of sourceEntries) {
-                    buckets.set(se.sourceIndex, [])
-                }
-                let seIdx = 0
-                for (const snap of allBatchSnapshots) {
-                    if (sourceEntries.length === 0) {
-                        break
-                    }
-                    while (seIdx < sourceEntries.length - 1 && snap.timestamp > sourceEntries[seIdx].endMs) {
-                        seIdx++
-                    }
-                    buckets.get(sourceEntries[seIdx].sourceIndex)?.push(snap)
-                }
-                for (const se of sourceEntries) {
-                    cache.store.markLoaded(se.sourceIndex, buckets.get(se.sourceIndex)!)
-                }
-
-                actions.loadNextSnapshotSource()
-                return
-            }
-
-            // Legacy path: snapshots are already in cache.snapshotsBySource from the loader
-            const sourceKey = snapshotsForSource.sources
-                ? keyForSource(snapshotsForSource.sources[0])
-                : keyForSource(snapshotsForSource.source)
-            const snapshotsData = (cache.snapshotsBySource || {})[sourceKey]
-            const snapshots = snapshotsData?.snapshots || []
-
-            if (!snapshots.length && sources?.length === 1 && sources[0].source !== SnapshotSourceType.file) {
-                posthog.capture('recording_snapshots_v2_empty_response', { source: sources[0] })
-            }
-
             actions.loadNextSnapshotSource()
         },
 
@@ -459,109 +424,33 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
                 return
             }
 
-            if (cache.useSnapshotStore) {
-                if (!cache.scheduler || !cache.store) {
-                    return
-                }
-                const batch = cache.scheduler.getNextBatch(cache.store, 10, cache.playbackPosition)
-                if (!batch) {
-                    actions.maybeStartPolling()
-                    return
-                }
-                const batchSources = batch.sourceIndices.map((i: number) => sources[i]).filter(Boolean)
-                if (batchSources.length > 0) {
-                    return actions.loadSnapshotsForSource(batchSources)
-                }
+            if (!cache.scheduler || !cache.store) {
+                return
+            }
+            const batch = cache.scheduler.getNextBatch(cache.store, 10, cache.playbackPosition)
+            if (!batch) {
                 actions.maybeStartPolling()
                 return
             }
-
-            // Legacy buffer_ahead loading path
-            const isSourceLoaded = (source: SessionRecordingSnapshotSource): boolean => {
-                const sourceKey = keyForSource(source)
-                return !!cache.snapshotsBySource?.[sourceKey]?.sourceLoaded
+            const batchSources = batch.sourceIndices.map((i: number) => sources[i]).filter(Boolean)
+            if (batchSources.length > 0) {
+                return actions.loadSnapshotsForSource(batchSources)
             }
-
-            const hasBlobV2 = sources.some((s) => s.source === SnapshotSourceType.blob_v2)
-
-            if (hasBlobV2) {
-                const nextSourcesToLoad = sources.filter(
-                    (s) => s.source !== SnapshotSourceType.file && !isSourceLoaded(s)
-                )
-
-                if (nextSourcesToLoad.length > 0) {
-                    return actions.loadSnapshotsForSource(nextSourcesToLoad.slice(0, 10))
-                }
-
-                actions.maybeStartPolling()
-            } else {
-                // V1 behavior unchanged
-                const nextSourceToLoad = sources.find((s) => {
-                    const sourceKey = keyForSource(s)
-                    return !cache.snapshotsBySource?.[sourceKey]?.sourceLoaded && s.source !== SnapshotSourceType.file
-                })
-
-                if (nextSourceToLoad) {
-                    return actions.loadSnapshotsForSource([nextSourceToLoad])
-                }
-            }
+            actions.maybeStartPolling()
         },
     })),
     selectors(({ cache }) => ({
         snapshotsLoading: [
-            (s) => [s.snapshotSourcesLoading, s.snapshotsForSourceLoading, s.snapshotsBySources, s.storeVersion],
-            (
-                snapshotSourcesLoading: boolean,
-                snapshotsForSourceLoading: boolean,
-                snapshotsBySources: Record<string, RecordingSnapshot[]>
-            ): boolean => {
-                if (cache.useSnapshotStore && cache.store) {
-                    return (
-                        cache.store.getAllLoadedSnapshots().length === 0 &&
-                        (snapshotSourcesLoading || snapshotsForSourceLoading)
-                    )
-                }
-                const snapshots = Object.values(snapshotsBySources).flat()
-                return snapshots?.length === 0 && (snapshotSourcesLoading || snapshotsForSourceLoading)
+            (s) => [s.snapshotSourcesLoading, s.snapshotsForSourceLoading, s.storeVersion],
+            (snapshotSourcesLoading: boolean, snapshotsForSourceLoading: boolean): boolean => {
+                return (
+                    (!cache.store || cache.store.getAllLoadedSnapshots().length === 0) &&
+                    (snapshotSourcesLoading || snapshotsForSourceLoading)
+                )
             },
         ],
 
         snapshotsLoaded: [(s) => [s.snapshotSources], (snapshotSources): boolean => !!snapshotSources],
-
-        snapshotsBySources: [
-            (s) => [s.snapshotsBySourceSuccessCount],
-            (
-                snapshotsBySourceSuccessCount: number
-            ): Record<SourceKey, SessionRecordingSnapshotSourceResponse> & { _count?: number } => {
-                // Store path doesn't use snapshotsBySources — return stable reference
-                // to avoid the shallow copy on every batch load.
-                if (cache.useSnapshotStore) {
-                    cache.stableEmptyBySources = cache.stableEmptyBySources || {}
-                    return cache.stableEmptyBySources
-                }
-
-                if (!cache.snapshotsBySource) {
-                    return {}
-                }
-
-                // KLUDGE: we keep the data in a cache so we can avoid creating large objects every time something changes
-                // KLUDGE: but if we change the data without changing the object instance then dependents don't recalculate
-                if (cache.snapshotsBySource['_count'] !== snapshotsBySourceSuccessCount) {
-                    // so we make a new object instance when the count changes
-                    // technically this should only be called when success count changes anyway...
-                    // but let's be very careful, it is relatively free to track the count
-                    // Create shallow copy to trigger dependent selectors
-                    // IMPORTANT: This must preserve the snapshot arrays from previous batches
-                    const newCache: Record<string, any> = {}
-                    for (const key of Object.keys(cache.snapshotsBySource)) {
-                        newCache[key] = cache.snapshotsBySource[key]
-                    }
-                    newCache['_count'] = snapshotsBySourceSuccessCount
-                    cache.snapshotsBySource = newCache
-                }
-                return cache.snapshotsBySource
-            },
-        ],
 
         isLoadingSnapshots: [
             (s) => [s.loadingSources],
@@ -576,13 +465,7 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
                 if (!snapshotSources || snapshotSources.length === 0) {
                     return false
                 }
-                if (cache.useSnapshotStore && cache.store) {
-                    return cache.store.allLoaded
-                }
-                return snapshotSources.every((source) => {
-                    const sourceKey = keyForSource(source)
-                    return cache.snapshotsBySource?.[sourceKey]?.sourceLoaded
-                })
+                return cache.store?.allLoaded ?? false
             },
         ],
 
@@ -610,7 +493,7 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
         isWaitingForPlayableFullSnapshot: [
             (s) => [s.storeVersion],
             (): boolean => {
-                if (!cache.useSnapshotStore || !cache.scheduler || !cache.store) {
+                if (!cache.scheduler || !cache.store) {
                     return false
                 }
                 const mode = cache.scheduler.currentMode
@@ -638,14 +521,11 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
             },
         ],
     })),
-    afterMount(({ actions, cache, values }) => {
+    afterMount(({ actions, cache }) => {
         cache.playerActive = true
-        const useStore = values.featureFlags[FEATURE_FLAGS.REPLAY_SNAPSHOT_STORE] === 'test'
-        cache.useSnapshotStore = useStore
-        if (useStore) {
-            cache.store = new SnapshotStore()
-            cache.scheduler = new LoadingScheduler()
-        }
+        cache.store = new SnapshotStore()
+        cache.scheduler = new LoadingScheduler()
+        cache.postProcessingState = createPostProcessingState()
 
         cache.disposables.add(() => {
             const handleVisibilityChange = (): void => {
@@ -667,11 +547,11 @@ export const snapshotDataLogic = kea<snapshotDataLogicType>([
         cache.playerActive = false
         cache.store = undefined
         cache.scheduler = undefined
-        cache.snapshotsBySource = undefined
-        cache.pendingBatch = undefined
         cache.previousSourceKeys = undefined
         cache.lastSourcesChangeTime = undefined
         cache.playbackPosition = undefined
         cache.loadFailureCount = undefined
+        cache.postProcessingState = undefined
+        cache.viewportForTimestamp = undefined
     }),
 ])
